requirements.txt
torch>=2.0
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0
matplotlib>=3.7
seaborn>=0.12
datasets>=2.14
tqdm>=4.66